Solvers refactoring + add g6 test choco + move source code for abscon and choco in respectively abscon.py and choco.py
