+++ solvers/solver.py
@@ -11,6 +11,93 @@ from pycsp3.classes.entities import VarEntities, EVarArray, EVar
 from pycsp3.tools.utilities import Stopwatch, flatten
 from pycsp3.dashboard import options
 
+def process_options(solving):
+    args = args_recursive = dict()
+    if solving[0] != '[':
+        assert "," not in solving and "]" not in solving
+        solver = solving
+    else:
+        assert solving[-1] == "]"
+        if "," not in solving:  # it means that only the name of the solver is between square brackets
+            solver = solving[1:-1]
+        else:
+            i = solving.find(",")
+            solver = solving[1:i]
+            args = option_parsing("[" + solving[i + 1:])
+            args_recursive = option_parsing("[" + solving[i + 1:], True) 
+    return solver, args, args_recursive
+
+def option_parsing(s, recursive=False):
+    if s is None:
+        return None
+    if s[0] != '[':
+        assert "," not in s and "]" not in s
+        return s
+    assert s[-1] == ']'
+    t = s[1:-1].split(",")
+    curr = -1
+    for i in range(len(t)):
+        if curr != -1:
+            t[curr] += "," + t[i]
+            if ']' in t[i]:
+                curr = -1
+            t[i] = ""
+        elif '[' in t[i] and ']' not in t[i]:
+            curr = i;
+    t = [v for v in t if v]  # we discard empty cells
+    t = [(k, None) if j == -1 else (k[:j], k[j + 1:]) for (j, k) in [(s.find("="), s) for s in t]]
+    return {k: v for (k, v) in [(k, option_parsing(v, recursive)) if recursive else (k, v) for (k, v) in t]}
+
+def simplify_args_recursive(args_recursive):
+    if "limit" in args_recursive:
+        def handle_limit(args_recursive, s):
+            if s.endswith("sols"):  # keep it at this position (because ends with s)
+                args_recursive["limit_sols"] = s[:-4]
+            elif s.endswith("runs"):
+                args_recursive["limit_runs"] = s[:-4]
+            elif s.endswith("h"):
+                args_recursive["limit_time"] = str(int(s[:-1]) * 3600)
+            elif s.endswith("m"):
+                args_recursive["limit_time"] = str(int(s[:-1]) * 60)
+            elif s.endswith("s"):
+                args_recursive["limit_time"] = s[:-1]
+            return args_recursive
+
+        v = args_recursive["limit"]
+        if isinstance(v, dict):
+            for key in v:
+                args_recursive = handle_limit(args_recursive, key)
+        else:
+            args_recursive = handle_limit(args_recursive, v)
+        del args_recursive["limit"]
+    if "restarts" in args_recursive:
+        v = args_recursive["restarts"]
+        if isinstance(v, dict):
+            for key in v:
+                if key in ["monotonic", "geometric", "luby"]:
+                    args_recursive["restarts_type"] = key
+                elif key == "cutoff":
+                    args_recursive["restarts_cutoff"] = v[key]
+                elif key == "factor":
+                    args_recursive["restarts_factor"] = v[key]
+                elif key == "gfactor":
+                    args_recursive["restarts_gfactor"] = v[key]
+        else:
+            args_recursive["restarts_type"] = v
+        del args_recursive["restarts"]
+    if "v" in args_recursive:
+        args_recursive["verbose"] = "1"
+        del args_recursive["v"]
+    if "vv" in args_recursive:
+        args_recursive["verbose"] = "2"
+        del args_recursive["vv"]
+    if "vvv" in args_recursive:
+        args_recursive["verbose"] = "3"
+        del args_recursive["vvv"]
+    return args_recursive
+
+
+
 class Instantiation:
     def __init__(self, pretty_solution, variables, values):
         self.pretty_solution = pretty_solution
